Update course controller method, general cleanup

+++ src/main/java/net/nsnsns/ciscms/repos/GradeableRepository.java
@@ -3,5 +3,5 @@ package net.nsnsns.ciscms.repos;
 import net.nsnsns.ciscms.models.Gradeable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GradeableRepository extends JpaRepository<Gradeable, Long> {
+public interface GradeableRepository extends JpaRepository<Gradeable, Integer> {
 }
